fix(ci): fix Python SDK ruff/mypy failures and lint job ordering
Python SDK (ruff):
- Auto-fix 70 ruff violations (UP/F/I/SIM rules): modernise imports,
  update type annotation syntax, remove unused imports, sort blocks
- Manually wrap 3 long lines (E501) in agentlens.py and patchers/openai.py
- Fix SpanAttributes TypeAlias to use Union[] (Python 3.9 safe)
- Cast transport._backoff return to float (mypy no-any-return)
- Remove unused # type: ignore[import] on openai/anthropic patchers
- Add eval_type_backport>=0.2.0; python_version<'3.10' dependency so
  Pydantic v2 can evaluate str|None annotations at runtime on Python 3.9

CI workflow:
- Change lint job to needs: build (was: install) so @agentlens/core/dist
  exists when ESLint type-checks sdk-anthropic; previously missing dist
  caused AgentLens to resolve as error type, triggering no-unsafe-member-access
- Add build outputs cache restore step to lint job

+++ packages/sdk-python/agentlens/transport.py
@@ -113,4 +113,4 @@ class Transport:
             except ValueError:
                 pass
         cap = _BASE_DELAY_S * (2**attempt)
-        return random.random() * cap  # full jitter
+        return float(random.random() * cap)  # full jitter
